refactor from commonjs to es6
