+++ package.json
@@ -20,5 +20,6 @@
     "knex": "^3.1.0",
     "mysql2": "^3.9.3",
     "sqlite3": "^5.1.7"
-  }
+  },
+  "type": "module"
 }
